requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
coreapi==2.3.3
python-dotenv==0.19.2
coreschema==0.0.4
python-dotenv==0.19.2
Django==3.2.5
django-viewflow==2.0.0a2
django-cors-headers==3.10.0
django-jazzmin==2.5.0
django-redis==5.2.0
django-material-admin==1.8.5
django-rest-swagger==2.2.0
django-tenants==3.4.0
djangorestframework==3.12.4
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
openapi-codec==1.3.2
promise==2.3
psycopg2-binary==2.8.6
pytz==2021.3
testresources==2.0.1
requests==2.26.0
Rx==1.6.1
simplejson==3.17.6
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
django-debug-toolbar==3.2.4
numpy==1.22.3
pandas==1.4.2
python-dateutil==2.8.2
